Complete domUpdates to make site fully functional
