+++ src/scripts.js
@@ -1,24 +1,22 @@
 // IMPORTS
 import './styles/index.scss';
+import RecipeRepository from '../src/classes/RecipeRepository';
+import Recipe from '../src/classes/Recipe';
+import User from '../src/classes/User';
+import Pantry from '../src/classes/Pantry';
 import {
   userAPI,
   ingredientAPI,
   recipeAPI,
   post
 } from './apiCalls';
-import RecipeRepository from '../src/classes/RecipeRepository';
-import Recipe from '../src/classes/Recipe';
-import User from '../src/classes/User';
-import Pantry from '../src/classes/Pantry';
 import {
   domUpdates,
-  displayRecipes,
-  show,
-  hide,
+  dropdownElement,
+  quantityInput
 } from './domUpdates';
 
-
-//API
+// GLOBAL VARIABLES
 let userIndex;
 let currentUser;
 let recipeList;
@@ -26,7 +24,10 @@ let recipeRepository;
 let usersData;
 let ingredientsData;
 let recipeData;
+let postMessage;
+let selectedRecipe;
 
+// FUNCTIONS
 Promise.all([userAPI, ingredientAPI, recipeAPI])
   .then(data => {
     [usersData, ingredientsData, recipeData] = [data[0], data[1], data[2]];
@@ -36,62 +37,11 @@ Promise.all([userAPI, ingredientAPI, recipeAPI])
       return new Recipe(recipe);
     })
     recipeRepository = new RecipeRepository(recipeList);
-    displayRecipes(recipeRepository.recipeData);
+    domUpdates.displayRecipes(recipeRepository.recipeData);
     domUpdates.populateDropdown();
   })
   .catch(error => console.log(error));
 
-// GLOBAL VARIABLES
-let recipeCards = [];
-let favoriteButtons = [];
-let selectedRecipe;
-let postMessage;
-
-// QUERY SELECTORS
-// const recipeSection = document.querySelector('.recipes-section-js');
-// const selectedRecipeView = document.querySelector('.individual-recipe-container-js');
-// const filterSection = document.querySelector('.filter-section-js');
-// const mainView = document.querySelector('.main-view-container-js');
-// const searchRecipesButton = document.getElementById('searchRecipes');
-// const filterButton = document.getElementById('filterButton');
-// const whatsCookin = document.querySelector('h1');
-// const searchBar = document.getElementById('searchInput');
-// const filterTags = document.querySelectorAll('.tag-js');
-// const favoritePageButton = document.getElementById('favoritesPage');
-// const addToCookbookButton = document.querySelector('.cookbook-button-js')
-// const cookbookPageButton = document.querySelector('.cookbook-page-button-js');
-// const singleViewFavoriteButton = document.querySelector('.single-view-favorite-button-js');
-// const pageTitle = document.querySelector('.page-title-js');
-// const pantryView = document.querySelector('.pantry-view-container-js');
-// const pantryPageButton = document.querySelector('.pantry-page-button-js');
-// const searchButton = document.getElementById('searchRecipes');
-// const highlightKey = document.querySelector('.key-js');
-// const dropdownElement = document.querySelector('.dropdown-js');
-// const quantityInput = document.querySelector('.quantity-input-js');
-// const addToPantryButton = document.querySelector('.form-button-js');
-
-
-// EVENT LISTENERS
-// searchRecipesButton.addEventListener('click', () => {
-//   searchAllRecipes(recipeRepository.recipeData);
-// });
-// searchBar.addEventListener('keyup', function (e) {
-//   if (e.keyCode === 13) {
-//     searchAllRecipes(recipeRepository.recipeData);
-//   }
-// })
-// filterButton.addEventListener('click', () => {
-//   filterRecipesByTag(recipeRepository.recipeData);
-// });
-// whatsCookin.addEventListener('click', displayHomePage);
-// favoritePageButton.addEventListener('click', domUpdates.displayFavorites);
-// addToCookbookButton.addEventListener('click', toggleCookbookButton);
-// cookbookPageButton.addEventListener('click', displayCookbook);
-// singleViewFavoriteButton.addEventListener('click', favoriteFromSingleRecipeView);
-// pantryPageButton.addEventListener('click', displayPantryView);
-// addToPantryButton.addEventListener('click', addIngredientToPantry);
-
-// FUNCTIONS
 async function postToPantry(info) {
   try {
     let response = await post(info)
@@ -102,23 +52,6 @@ async function postToPantry(info) {
     console.log(error)
    }
 }
-
-// function displayMessageDropdown() {
-//   const dropdownMessage = document.querySelector('.dropdown-message-js');
-//   dropdownMessage.innerText = `${postMessage}`;
-//   dropdownMessage.classList.remove('hidden-visibility')
-//   setTimeout(() => {dropdownMessage.classList.add('fade-out')}, 2000);
-//   setTimeout(() => {
-//     dropdownMessage.classList.add('hidden-visibility');
-//     dropdownMessage.classList.remove('fade-out');
-//   }, 4000)
-// }
-
-// function displayMessageButtons(e) {
-//   const buttonMessage = document.getElementById(`m${e.target.id}`);
-//   buttonMessage.innerText = postMessage;
-//   setTimeout(() => {buttonMessage.classList.add('fade-out')}, 2000);
-// }
 
 async function getPantry() {
   let response = await fetch("http://localhost:3001/api/v1/users")
@@ -138,272 +71,9 @@ async function addIngredientToPantry() {
   domUpdates.clearInputs()
 }
 
-// function clearInputs() {
-//   dropdownElement.value = 'choose ingredient';
-//   quantityInput.value = '';
-// }
-
-function reassignUserPantry(data) {
-  const updatedUser = data.find((user) => {
-    return user.id === currentUser.id
-  });
-  currentUser.pantry = new Pantry(updatedUser.pantry)
-}
-
-// function displayFavorites() {
-//   whatsCookin.classList.remove('home-page');
-//   hide([favoritePageButton, pantryView, highlightKey, selectedRecipeView]);
-//   show([filterSection, recipeSection, mainView, cookbookPageButton, pageTitle, pantryPageButton]);
-//   pageTitle.innerText = 'my favorites';
-//   filterButton.addEventListener('click', () => {
-//     filterRecipesByTag(currentUser.favorites)
-//   });
-//   searchRecipesButton.addEventListener('click', () => {
-//     searchAllRecipes(currentUser.favorites)
-//   });
-//   displayRecipes(currentUser.favorites);
-//   favoriteButtons = document.querySelectorAll('.favorite-button-js');
-//   favoriteButtons.forEach((button) => {
-//     button.addEventListener('click', removeFromPage);
-//   })
-// }
-
-// function removeFromPage() {
-//   currentUser.removeFromFavorites();
-//   domUpdates.displayFavorites();
-// }
-
-// function toggleFavoriteButton(e) {
-//   if (e.target.classList.contains('favorite-button-js')) {
-//     const recipeID = Number(e.target.closest('section').id.slice(2));
-//     const recipe = recipeRepository.recipeData.find((element) => {
-//       return element.id === recipeID;
-//     });
-//     if (e.target.value === 'unfavorited') {
-//       selectFavoriteButton(e);
-//       currentUser.addToFavorites(recipe);
-//     } else if (e.target.value === 'favorited') {
-//       deselectFavoriteButton(e);
-//       currentUser.removeFromFavorites(recipe);
-//     }
-//   }
-// }
-
-// function selectFavoriteButton(e) {
-//   e.target.value = 'favorited';
-//   e.target.classList.add('favorited-state');
-// }
-//
-// function deselectFavoriteButton(e) {
-//   e.target.value = 'unfavorited';
-//   e.target.classList.remove('favorited-state');
-// }
-
-// function favoriteFromSingleRecipeView() {
-//   if (currentUser.favorites.includes(selectedRecipe)) {
-//     singleViewFavoriteButton.value = 'unfavorited';
-//     singleViewFavoriteButton.classList.remove('favorited-state');
-//     currentUser.removeFromFavorites(selectedRecipe);
-//   } else {
-//     singleViewFavoriteButton.value = 'unfavorited';
-//     singleViewFavoriteButton.classList.add('favorited-state');
-//     currentUser.addToFavorites(selectedRecipe);
-//   }
-// }
-
-// function displayCookbook() {
-//   whatsCookin.classList.remove('home-page');
-//   hide([addToCookbookButton, selectedRecipeView, cookbookPageButton, pantryView]);
-//   show([filterSection, mainView, recipeSection, favoritePageButton, addToCookbookButton, pageTitle, highlightKey, pantryPageButton]);
-//   pageTitle.innerText = 'my cookbook';
-//   filterButton.addEventListener('click', () => {
-//     filterRecipesByTag(currentUser.recipesToCook);
-//   });
-//   searchRecipesButton.addEventListener('click', () => {
-//     searchAllRecipes(currentUser.recipesToCook);
-//   });
-//   displayRecipes(currentUser.recipesToCook);
-//   showRecipesUserCanCook();
-// }
-
-// function showRecipesUserCanCook() {
-//   const recipeCards = document.querySelectorAll('.recipe-card-js');
-//   const recipesUserCanCook = recipeCards.forEach(recipeCard => {
-//     const recipeOnCard = currentUser.recipesToCook.find(recipe => {
-//       return recipe.id === parseInt(recipeCard.id.slice(2))
-//     })
-//
-//     const haveIngredients = currentUser.pantry.checkPantry(recipeOnCard);
-//     if (haveIngredients) {
-//       recipeCard.childNodes[3].classList.add('highlight-card');
-//     }
-//   })
-//
-//   return recipesUserCanCook;
-// }
-
-// function toggleCookbookButton() {
-//   if (addToCookbookButton.value === 'false') {
-//     selectAddToCookbookButton();
-//     currentUser.addToRecipesToCook(selectedRecipe);
-//     addToCookbookButton.innerText = 'remove from cookbook';
-//   } else if (addToCookbookButton.value === 'true') {
-//     deselectAddToCookbookButton();
-//     currentUser.removeFromRecipesToCook(selectedRecipe);
-//     addToCookbookButton.innerText = 'add to cookbook';
-//   }
-// }
-
-// function selectAddToCookbookButton() {
-//   addToCookbookButton.value = 'true';
-//   addToCookbookButton.classList.add('in-cookbook-state');
-// }
-//
-// function deselectAddToCookbookButton() {
-//   addToCookbookButton.value = 'false';
-//   addToCookbookButton.classList.remove('in-cookbook-state');
-// }
-
-function getRandomIndex(array) {
-  return Math.floor(Math.random() * array.length);
-}
-
-function displayHomePage() {
-  if (!whatsCookin.classList.contains('home-page')) {
-    whatsCookin.classList.add('home-page');
-    displayRecipes(recipeRepository.recipeData);
-    searchBar.value = '';
-    hide([selectedRecipeView, pantryView, highlightKey]);
-    show([mainView, recipeSection, searchBar, searchRecipesButton, favoritePageButton, cookbookPageButton, pageTitle, pantryPageButton]);
-    pageTitle.innerText = 'home';
-    filterButton.addEventListener('click', () => {
-      filterRecipesByTag(recipeRepository.recipeData);
-    })
-    searchRecipesButton.addEventListener('click', () => {
-      searchAllRecipes(recipeRepository.recipeData);
-    })
-  }
-}
-
-function filterRecipesByTag(recipes) {
-  const selectedTags = [];
-  filterTags.forEach((tag) => {
-    if (tag.checked) {
-      selectedTags.push(tag.id);
-    }
-  })
-
-  if (selectedTags.length !== 0) {
-    recipeRepository.filterByTags(selectedTags, recipes);
-    displayRecipes(recipeRepository.recipesToShow);
-  } else {
-    displayRecipes(recipes);
-  }
-}
-
-// function displayRecipes(recipes) {
-//   recipeSection.innerHTML = '';
-//   recipes.forEach(recipe => {
-//     recipeSection.innerHTML += `
-//       <section class='flex column align-center recipe-card recipe-card-js' id='id${recipe.id}'>
-//          <img class='recipe-card-image' src=${recipe.image} alt='recipe image' class='recipe-photo'>
-//          <div class='flex column align-center recipe-card-text'>
-//            <button class='favorite-button favorite-button-js' value='unfavorited'>favorite</button>
-//            <p class='recipe-card-name'>${recipe.name}</p>
-//           </div>
-//        </section>
-//      `;
-//   })
-//
-//   recipeCards = document.querySelectorAll('.recipe-card-js');
-//   addCardInfo(recipeCards);
-//
-//   favoriteButtons = document.querySelectorAll('.favorite-button-js');
-//   updateFavoriteButton(favoriteButtons);
-// }
-
-// function updateFavoriteButton(favoriteButtons) {
-//   favoriteButtons.forEach((button) => {
-//     button.addEventListener('click', function (e) {
-//       toggleFavoriteButton(e);
-//     })
-//   })
-// }
-
-// function addCardInfo(recipeCards) {
-//   recipeCards.forEach((card, index) => {
-//     const cardId = Number(card.id.slice(2));
-//     const currentRecipe = recipeRepository.recipeData.find(recipe => {
-//       return recipe.id === cardId;
-//     })
-//
-//     const button = card.childNodes[3].childNodes[1];
-//     if (currentUser.favorites.includes(currentRecipe)) {
-//       button.value = 'favorited';
-//       button.classList.add('favorited-state');
-//     }
-//     card.addEventListener('click', function (e) {
-//       displaySelectedRecipe(e);
-//     })
-//   })
-// }
-
-// function searchAllRecipes(recipes) {
-//   const searchName = document.getElementById('searchInput').value;
-//   recipeRepository.filterByNameOrIng(searchName, ingredientsData, recipes);
-//   displayRecipes(recipeRepository.recipesToShow);
-// }
-
-function displaySelectedRecipe(e) {
-  if (!e.target.classList.contains('favorite-button-js')) {
-    whatsCookin.classList.remove('home-page');
-    const recipeID = Number(e.target.closest('section').id.slice(2));
-    selectedRecipe = recipeRepository.recipeData.find((currentRecipe) => {
-      return currentRecipe.id === recipeID;
-    })
-    show([selectedRecipeView, favoritePageButton, cookbookPageButton, pantryPageButton]);
-    hide([mainView, searchBar, searchButton, pageTitle, pantryView, highlightKey]);
-
-    showCookbookStatus(selectedRecipe);
-    showFavoritesStatus(selectedRecipe);
-    updateRecipeText(e, selectedRecipe, ingredientsData);
-    displayIngredientsNeeded(selectedRecipe);
-  }
-}
-
-function displayIngredientsNeeded(recipe) {
-  const neededIngredients = currentUser.pantry.getMissingIngredients(recipe);
-  const neededIngredientsSection = document.querySelector('.ingredients-needed-js');
-
-  if (neededIngredients.length === 0) {
-    neededIngredientsSection.innerHTML =
-      `<p>you have all of the ingredients needed to cook this recipe!</p>
-      <button class="cook-recipe-js">cook recipe</button>`;
-
-    const cookRecipeButton = document.querySelector('.cook-recipe-js');
-    cookRecipeButton.addEventListener('click', removeIngredients)
-  } else {
-    const elements = neededIngredients.reduce((acc, ingredient) => {
-      const matchedId = ingredientsData.find(item => {
-        return item.id === ingredient.ingredient
-      })
-      const ingWithUnits = selectedRecipe.ingredients.find(item => {
-        return item.id === ingredient.ingredient
-      })
-
-      acc += `
-        ${matchedId.name}: ${ingredient.amount} ${ingWithUnits.quantity.unit}<br>
-      `
-      return acc;
-    }, '<h3>for this recipe you are missing: </h3>')
-    neededIngredientsSection.innerHTML = elements;
-  }
-}
-
 async function removeRecipeIngredients() {
   for (const ingredient of selectedRecipe.ingredients) {
     const amount = -(parseFloat(ingredient.quantity.amount));
-    // await postToPantry(ingredient.id, amount);
     const info = currentUser.modifyIngredient(ingredient.id, amount);
     await postToPantry(info);
   }
@@ -414,56 +84,6 @@ async function removeIngredients() {
   await getPantry();
   displayIngredientsNeeded(selectedRecipe);
 }
-
-function displayPantryView() {
-  if (pantryView.classList.contains('hidden')) {
-    whatsCookin.classList.remove('home-page');
-    pageTitle.innerText = "my pantry";
-    show([pantryView, cookbookPageButton, favoritePageButton]);
-    hide([mainView, searchBar, searchButton, pantryPageButton, selectedRecipeView]);
-    domUpdates.populatePantry();
-  }
-}
-
-// function populatePantry() {
-//   createTable();
-//   selectTableButtons();
-// }
-
-// function selectTableButtons() {
-//   const plusButtons = document.querySelectorAll('.plus-js');
-//   const minusButtons = document.querySelectorAll('.minus-js');
-//   plusButtons.forEach(button => {
-//     button.addEventListener('click', function (e) {
-//       changeAmount(e)
-//     });
-//   })
-//   minusButtons.forEach(button => {
-//     button.addEventListener('click', function (e) {
-//       changeAmount(e)
-//     });
-//   })
-// }
-
-// function createTable() {
-//   const tableBody = document.querySelector('tbody')
-//   tableBody.innerHTML = '';
-//   currentUser.pantry.ingredients.forEach((item) => {
-//     const ingredientData = ingredientsData.find((ingredient) => {
-//       return ingredient.id === item.ingredient
-//     })
-//     tableBody.innerHTML += `
-//     <tr>
-//     <td>${ingredientData.name}</td>
-//     <td>${item.amount}</td>
-//     <td class="button-column flex">
-//     <button class="round-buttons minus-js" id="${ingredientData.id}">-</button>
-//     <button class="round-buttons plus-js" id="${ingredientData.id}">+</button>
-//     <p id='m${ingredientData.id}' class ='button-message'></p>
-//     </td>
-//     </tr>`
-//   });
-// }
 
 async function changeAmount(e) {
   const currentAmount = currentUser.pantry.ingredients.find(item => {
@@ -485,82 +105,68 @@ async function changeAmount(e) {
   domUpdates.displayMessageButtons(e, postMessage);
 }
 
-function populateDropdown() {
-  ingredientsData.sort((a, b) => {
-    return a.name.localeCompare(b.name, 'en', {
-      ignorePunctuation: true
-    });
-  }).forEach((ingredient) => {
-    if (ingredient.name) {
-      dropdownElement.innerHTML += `<option value="${ingredient.id}">${ingredient.name}</option>"`
-    }
+function reassignUserPantry(data) {
+  const updatedUser = data.find((user) => {
+    return user.id === currentUser.id
   });
+  currentUser.pantry = new Pantry(updatedUser.pantry)
 }
 
-function showCookbookStatus(selectedRecipe) {
-  if (currentUser.recipesToCook.includes(selectedRecipe)) {
-    selectAddToCookbookButton();
-    addToCookbookButton.innerText = 'remove from cookbook';
-  } else {
-    deselectAddToCookbookButton()
-    addToCookbookButton.innerText = 'add to cookbook';
-  }
-}
-
-function showFavoritesStatus(selectedRecipe) {
-  if (currentUser.favorites.includes(selectedRecipe)) {
-    singleViewFavoriteButton.value = 'favorited';
-    singleViewFavoriteButton.classList.add('favorited-state');
-  } else {
-    singleViewFavoriteButton.value = 'unfavorited';
-    singleViewFavoriteButton.classList.remove('favorited-state');
-  }
-}
-
-function updateRecipeText(e, selectedRecipe, ingredientsData) {
-  const instructionsSection = document.querySelector('.instructions-section-js');
-  const ingredientListSection = document.querySelector('.ingredient-list-section-js');
-  const image = document.querySelector('.selected-recipe-photo-js');
-  const name = document.querySelector('.selected-recipe-name-js');
-  const cost = document.querySelector('.cost-js');
-
-  const ingredientListElement = getIngredientListElement(e, selectedRecipe);
-  const instructionsElement = getInstructionsElement(e, selectedRecipe);
-  const costNum = selectedRecipe.calculateRecipeCostInDollars(ingredientsData);
-
-  ingredientListSection.innerHTML += ingredientListElement;
-  instructionsSection.innerHTML += instructionsElement;
-  image.src = selectedRecipe.image;
-  name.innerText = selectedRecipe.name;
-  cost.innerText = `$${costNum}`;
-}
-
-function getIngredientListElement(e, selectedRecipe) {
-  const ingredientListSection = document.querySelector('.ingredient-list-section-js');
-  ingredientListSection.innerHTML = '<h3>ingredients</h3>';
-  const ingredientNames = selectedRecipe.determineRecipeIngredients(ingredientsData);
-  const ingredientListText = selectedRecipe.ingredients.reduce((acc, ingredient, index) => {
-    let amount = ingredient.quantity.amount
-    if (!(ingredient.quantity.amount % 1) === 0) {
-      amount = ingredient.quantity.amount.toFixed(2)
+function filterRecipesByTag(recipes) {
+  const selectedTags = [];
+  filterTags.forEach((tag) => {
+    if (tag.checked) {
+      selectedTags.push(tag.id);
     }
-    acc += `${ingredientNames[index]}: ${amount} ${ingredient.quantity.unit}<br><br>`;
-    return acc;
-  }, '');
+  })
 
-  return '<p class="selected-recipe-ingredients-list">' + ingredientListText + '</p>';
+  if (selectedTags.length !== 0) {
+    recipeRepository.filterByTags(selectedTags, recipes);
+    displayRecipes(recipeRepository.recipesToShow);
+  } else {
+    displayRecipes(recipes);
+  }
 }
 
-function getInstructionsElement(e, selectedRecipe) {
-  const instructionsSection = document.querySelector('.instructions-section-js');
-  instructionsSection.innerHTML = '<h3>instructions</h3>';
-  const instructionsStrings = selectedRecipe.returnInstructions();
-  const instructionsText = instructionsStrings.reduce((acc, instruction) => {
-    acc += `${instruction}<br><br>`;
-    return acc;
-  }, '');
+function addCardInfo(recipeCards) {
+  recipeCards.forEach((card, index) => {
+    const cardId = Number(card.id.slice(2));
+    const currentRecipe = recipeRepository.recipeData.find(recipe => {
+      return recipe.id === cardId;
+    })
 
-  return '<p class="selected-recipe-instructions">' + instructionsText + '</p>'
+    const button = card.childNodes[3].childNodes[1];
+    if (currentUser.favorites.includes(currentRecipe)) {
+      button.value = 'favorited';
+      button.classList.add('favorited-state');
+    }
+    card.addEventListener('click', function (e) {
+      domUpdates.displaySelectedRecipe(e);
+    })
+  })
 }
 
-export {currentUser}
+function updateSelectedRecipe(e) {
+  const recipeID = Number(e.target.closest('section').id.slice(2));
+  selectedRecipe = recipeRepository.recipeData.find((currentRecipe) => {
+    return currentRecipe.id === recipeID;
+  })
+}
+
+function getRandomIndex(array) {
+  return Math.floor(Math.random() * array.length);
+}
+
+export {
+  recipeRepository,
+  ingredientsData,
+  currentUser,
+  selectedRecipe,
+  postMessage,
+  filterRecipesByTag,
+  addIngredientToPantry,
+  addCardInfo,
+  changeAmount,
+  removeIngredients,
+  updateSelectedRecipe
+}
